Use linear interpolation when resampling

+++ src/audio/resample.h
@@ -1,9 +1,9 @@
 #include "../playroutine/play_mod.h"
 
 const static int PITCH_QUANTA = 2047;
-const static double PHASE_INCREMENT_CONVERSION = 3273808.59375;
+const static float PHASE_INCREMENT_CONVERSION = 3273808.59375f;
 
-static double *phase_increments;
+static float *phase_increments;
 static float *resample_buffer;
 static size_t resample_buffer_bytes;
